move pillars to lib
