+++ frontend/web/components/Header/Nav/Pillars/index.tsx
@@ -9,7 +9,7 @@ import {
 } from '@guardian/pasteup/breakpoints';
 
 import { serif } from '@guardian/pasteup/fonts';
-import { pillarMap, pillarPalette } from '../../../../../pillars';
+import { pillarMap, pillarPalette } from '../../../../../lib/pillars';
 import { palette } from '@guardian/pasteup/palette';
 
 const pillarColours = pillarMap(
